fix proof of attestation cron
