+++ packages/brain/src/modules/cron/index.ts
@@ -1,3 +1,3 @@
 export { CronJob } from "./cron.js";
 export { ReloadValidators } from "./reloadValidators.js";
-export { ProofOfAttestation } from "./proofOfAttestation.js";
+export { sendProofOfAttestation } from "./sendProofOfAttestation.js";
